10.07 update

+++ Resource.Web/Controllers/UserController.cs
@@ -25,11 +25,11 @@ namespace Resource.Web.Controllers
             var list = dc.Set<T_User>()
                 .AsNoTracking()
                 .Where(a => true)
-                .Select(a => new { a.Account, a.UserName, a.CreateDate, a.Addr, a.Enable, a.Email, a.Phone });
+                .Select(a => new { a.Account, a.UserName, a.CreateTime, a.Addr, a.Enable, a.Email, a.Phone });
             if (!string.IsNullOrEmpty(param.ID)) list = list.Where(a => a.Account.Contains(param.ID));
             if (!string.IsNullOrEmpty(param.Name)) list = list.Where(a => a.UserName.Contains(param.Name));
-            if (param.Stime != null) list = list.Where(a => a.CreateDate >= param.Stime);
-            if (param.Etime != null) list = list.Where(a => a.CreateDate <= param.Etime);
+            if (param.Stime != null) list = list.Where(a => a.CreateTime >= param.Stime);
+            if (param.Etime != null) list = list.Where(a => a.CreateTime <= param.Etime);
             if (param.Enable != null) list = list.Where(a => a.Enable == param.Enable);
             int count = list.Count();
             list = list.OrderBy(a => a.Account)
@@ -74,14 +74,6 @@ namespace Resource.Web.Controllers
                     ud.DataID = item;
                     dc.Set<T_UserData>().Add(ud);
                 }
-                //绑定角色权限
-                T_UserRole role = new T_UserRole();
-                role.UserID = newUser.Account;
-                if (!string.IsNullOrEmpty(form["Role"]) && form["Role"] == "1")
-                    role.RoleID = 1;
-                else
-                    role.RoleID = 2;
-                dc.Set<T_UserRole>().Add(role);
                 dc.SaveChanges();
                 return Json(Result.Success());
             }
@@ -103,19 +95,30 @@ namespace Resource.Web.Controllers
             {
                 T_User updateUser = dc.Set<T_User>().Where(a => a.Account.Equals(id)).FirstOrDefault();
                 if (updateUser == null) return Json(Result.Fail(msg: "系统无此用户，更新失败！"));
-                if (!TryUpdateModel(updateUser, "", form.AllKeys, new string[] { "Enable", "PWD", "CreateTime", "CreateUser" }))
-                    return Json(Result.Fail());
-                updateUser.UpdateTime = DateTime.Now;
-                updateUser.UpdateUser = user.Account;
-                //绑定数据权限
-                dc.Database.ExecuteSqlCommand(string.Format("delete from t_userdata where userid='{0}'", updateUser.Account));
-                foreach (var item in form["Park"].Split(','))
+                if (updateUser.Account == "admin")
                 {
-                    T_UserData ud = new T_UserData();
-                    ud.UserID = updateUser.Account;
-                    ud.DataID = item;
-                    dc.Set<T_UserData>().Add(ud);
+                    if (!TryUpdateModel(updateUser, "", form.AllKeys, new string[] { "Enable", "PWD", "CreateTime", "CreateUser", "RoleID" }))
+                        return Json(Result.Fail());
+                    updateUser.UpdateTime = DateTime.Now;
+                    updateUser.UpdateUser = user.Account;
                 }
+                else
+                {
+                    if (!TryUpdateModel(updateUser, "", form.AllKeys, new string[] { "Enable", "PWD", "CreateTime", "CreateUser" }))
+                        return Json(Result.Fail());
+                    updateUser.UpdateTime = DateTime.Now;
+                    updateUser.UpdateUser = user.Account;
+                    //绑定数据权限
+                    dc.Database.ExecuteSqlCommand(string.Format("delete from t_userdata where userid='{0}'", updateUser.Account));
+                    foreach (var item in form["Park"].Split(','))
+                    {
+                        T_UserData ud = new T_UserData();
+                        ud.UserID = updateUser.Account;
+                        ud.DataID = item;
+                        dc.Set<T_UserData>().Add(ud);
+                    }
+                }
+
                 dc.SaveChanges();
                 return Json(Result.Success());
             }
@@ -129,7 +132,6 @@ namespace Resource.Web.Controllers
             T_User user = dc.Set<T_User>().Where(a => a.Account == id).FirstOrDefault();
             return View(user);
         }
-
         public ActionResult SingleReset()
         {
             return View(user);
@@ -156,7 +158,6 @@ namespace Resource.Web.Controllers
                 return Json(Result.Exception(exmsg: ex.StackTrace));
             }
         }
-
         [HttpPost]
         public JsonResult Reset(string id, FormCollection form)
         {
@@ -174,7 +175,10 @@ namespace Resource.Web.Controllers
                 return Json(Result.Exception(exmsg: ex.StackTrace));
             }
         }
-        public ActionResult Role(string id)
+
+        #region 多重绑定角色代码
+        /*
+		  public ActionResult Role(string id)
         {
             ViewBag.Role = dc.Set<T_Role>().ToList();
             T_User user = dc.Set<T_User>().Where(a => a.Account == id).FirstOrDefault();
@@ -194,33 +198,37 @@ namespace Resource.Web.Controllers
                 {
                     dc.Set<T_UserRole>().Add(item);
                 }
-                if (dc.SaveChanges() > 0) return Json(Result.Success());
-                return Json(Result.Fail());
+             dc.SaveChanges()；
+                return Json(Result.Success());
             }
             catch (Exception ex)
             {
                 return Json(Result.Exception(exmsg: ex.StackTrace));
             }
         }
+        */
+        #endregion
+
         [HttpPost]
         public JsonResult Del(string id)
         {
             try
             {
-                T_User user = dc.Set<T_User>().Where(a => a.Account == id).FirstOrDefault();
-                var ur = user.T_UserRole.ToList();
-                foreach (var item in ur)
-                {
-                    dc.Set<T_UserRole>().Remove(item);
-                }
-                var li = user.T_LoginInfo.ToList();
+                T_User updateUser = dc.Set<T_User>().Where(a => a.Account == id).FirstOrDefault();
+                if (updateUser.Account == "admin") return Json(Result.Fail(msg: "无法操作超级管理员！"));
+                //var ur = updateUser.T_UserRole.ToList();
+                //foreach (var item in ur)
+                //{
+                //    dc.Set<T_UserRole>().Remove(item);
+                //}
+                var li = updateUser.T_LoginInfo.ToList();
                 foreach (var item in li)
                 {
                     dc.Set<T_LoginInfo>().Remove(item);
                 }
-                dc.Set<T_User>().Remove(user);
-                if (dc.SaveChanges() > 0) return Json(Result.Success());
-                return Json(Result.Fail());
+                dc.Set<T_User>().Remove(updateUser);
+                dc.SaveChanges();
+                return Json(Result.Success());
             }
             catch (Exception ex)
             {
@@ -232,12 +240,13 @@ namespace Resource.Web.Controllers
             try
             {
                 T_User updateUser = dc.Set<T_User>().Where(a => a.Account == id).FirstOrDefault();
+                if (updateUser.Account == "admin") return Json(Result.Fail(msg: "无法操作超级管理员！"));
                 updateUser.Enable = true;
                 updateUser.UpdateTime = DateTime.Now;
                 updateUser.UpdateUser = user.Account;
                 dc.Set<T_User>().AddOrUpdate(updateUser);
-                if (dc.SaveChanges() > 0) return Json(Result.Success());
-                return Json(Result.Fail());
+                dc.SaveChanges();
+                return Json(Result.Success());
             }
             catch (Exception ex)
             {
@@ -249,31 +258,54 @@ namespace Resource.Web.Controllers
             try
             {
                 T_User updateUser = dc.Set<T_User>().Where(a => a.Account == id).FirstOrDefault();
+                if (updateUser.Account == "admin") return Json(Result.Fail(msg: "无法操作超级管理员！"));
                 updateUser.Enable = false;
                 updateUser.UpdateTime = DateTime.Now;
                 updateUser.UpdateUser = user.Account;
                 dc.Set<T_User>().AddOrUpdate(updateUser);
-                if (dc.SaveChanges() > 0) return Json(Result.Success());
-                return Json(Result.Fail());
+                dc.SaveChanges();
+                return Json(Result.Success());
             }
             catch (Exception ex)
             {
                 return Json(Result.Exception(exmsg: ex.StackTrace));
             }
         }
-
         public void SetPark(T_User cust)
         {
             List<T_UserData> custPark = null;
-            if (cust != null)
+            if ((cust != null && cust.Account == "admin") || user.Account == "admin")
+            {
+                var mastPark = dc.Set<T_Park>().Select(a => new { a.ID, a.Name, Enable = true }).ToList();
+                ViewBag.Park = JsonConvert.DeserializeObject(JsonConvert.SerializeObject(mastPark));
+            }
+            else if (cust != null)
+            {
                 custPark = cust.T_UserData.ToList();
-            if (custPark != null && custPark.Count() > 0)
+                if (custPark.Count() > 0)
+                {
+                    var mastPark = user.T_UserData
+                    .Select(a => new
+                    {
+                        a.DataID,
+                        Enable = custPark.Where(b => b.DataID == a.DataID).Count() > 0
+                    })
+                    .Join(dc.Set<T_Park>(), a => a.DataID, b => b.ID, (a, b) => new
+                    {
+                        ID = a.DataID,
+                        Name = b.Name,
+                        Enable = a.Enable
+                    }).ToList();
+                    ViewBag.Park = JsonConvert.DeserializeObject(JsonConvert.SerializeObject(mastPark));
+                }
+            }
+            else
             {
                 var mastPark = user.T_UserData
                 .Select(a => new
                 {
                     a.DataID,
-                    Enable = custPark.Where(b => b.DataID == a.DataID).Count() > 0
+                    Enable = false
                 })
                 .Join(dc.Set<T_Park>(), a => a.DataID, b => b.ID, (a, b) => new
                 {
@@ -283,23 +315,6 @@ namespace Resource.Web.Controllers
                 }).ToList();
                 ViewBag.Park = JsonConvert.DeserializeObject(JsonConvert.SerializeObject(mastPark));
             }
-            else
-            {
-                var mastPark = user.T_UserData
-                 .Select(a => new
-                 {
-                     a.DataID,
-                     Enable = false
-                 })
-                 .Join(dc.Set<T_Park>(), a => a.DataID, b => b.ID, (a, b) => new
-                 {
-                     ID = a.DataID,
-                     Name = b.Name,
-                     Enable = a.Enable
-                 }).ToList();
-                ViewBag.Park = JsonConvert.DeserializeObject(JsonConvert.SerializeObject(mastPark));
-            }
-
         }
     }
 }
